+++ demisto_sdk/__main__.py
@@ -3604,32 +3604,6 @@ def pre_commit(
     )
 
 
-@main.command(short_help="Run unit tests in a docker for integrations and scripts")
-@click.help_option("-h", "--help")
-@click.option(
-    "-i",
-    "--input",
-    type=PathsParamType(
-        exists=True, resolve_path=True
-    ),  # PathsParamType allows passing a list of paths
-    help="The path of the content pack/file to validate specifically.",
-)
-@click.option(
-    "-v", "--verbose", is_flag=True, default=False, help="Verbose output of unit tests"
-)
-@click.argument("file_paths", nargs=-1, type=click.Path(exists=True, resolve_path=True))
-@click.pass_context
-@logging_setup_decorator
-def run_unit_tests(
-    ctx, input: str, file_paths: Tuple[str, ...], verbose: bool, **kwargs
-):
-    if input:
-        file_paths = tuple(input.split(","))
-    from demisto_sdk.commands.run_unit_tests.unit_tests_runner import unit_test_runner
-
-    sys.exit(unit_test_runner(file_paths, verbose))
-
-
 @main.command(short_help="Setup integration environments")
 @click.option(
     "-i",